added if else logic to team form
