+++ scripts/landing page/setupGamePage.js
@@ -1,4 +1,4 @@
-import { TeamForm } from "../Forms/teamForm.js";
+import { TeamForm } from "../team/teamForm.js";
 import { playerForm } from "../player/createPlayerForm.js";
 const eventHub=document.querySelector(".container")
